requirements.txt
datasets==2.20.0
einops==0.8.0
flash-attn==2.6.1
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
supervision==0.22.0
timm==1.0.7
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.4
PyYAML==6.0
